requirement.txt
streamlit
openai
langchain
pypdf
faiss-cpu
tiktoken
python-dotenv
PyPDF2
fpdf